apps/api/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
alembic>=1.14
pydantic-settings>=2.7
pgvector>=0.3
celery[redis]>=5.4
redis>=5.2
openai>=1.58
pypdf>=5.1
httpx>=0.28
beautifulsoup4>=4.13
langchain-text-splitters>=0.3
langgraph>=0.2
python-jose[cryptography]>=3.3
bcrypt>=4.0
python-multipart>=0.0.20
aiofiles>=25.1.0
aioboto3>=15.5.0
authlib>=1.3
croniter>=6.0
feedparser>=6.0
aiosmtplib>=3.0
markdown>=3.5
